Add files via upload
Problem Statement: Given an undirected connected graph with V vertices and adjacency list adj. You are required to find all the vertices removing which (and edges through it) disconnect the graph into 2 or more components.
Articulation Points of a graph are the nodes on whose removal, the graph breaks into multiple components.